+++ projet/controllers/detail_controller.py
@@ -1,9 +1,17 @@
 from sqlalchemy.orm import Session
+from fastapi import HTTPException, status
 from ..models import Detail
 
 
 def get_detail(db: Session, detail_id: int):
-    return db.query(Detail).get(detail_id)
+    detail = db.query(Detail).get(detail_id)
+    if not detail:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Detail not found")
+    return detail
+
+
+def get_all_details(db: Session):
+    return db.query(Detail).all()
 
 
 def create_detail(db: Session, detail: Detail):
@@ -13,15 +21,22 @@ def create_detail(db: Session, detail: Detail):
     return detail
 
 
-def update_detail(db: Session, detail: Detail, updated_detail: Detail):
-    for attr, value in updated_detail.dict().items():
+def update_detail(db: Session, detail_id: int, updated_detail: Detail):
+    detail = db.query(Detail).get(detail_id)
+    if not detail:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Detail not found")
+    updated_data = updated_detail.dict(exclude_unset=True)
+    for attr, value in updated_data.items():
         setattr(detail, attr, value)
     db.commit()
     db.refresh(detail)
     return detail
 
 
-def delete_detail(db: Session, detail: Detail):
+def delete_detail(db: Session, detail_id: int):
+    detail = db.query(Detail).get(detail_id)
+    if not detail:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Detail not found")
     db.delete(detail)
     db.commit()
     return {"message": "Detail deleted"}
