+++ projet/controllers/detail_objet_controller.py
@@ -1,9 +1,17 @@
 from sqlalchemy.orm import Session
+from fastapi import HTTPException, status
 from ..models import DetailObjet
 
 
 def get_detail_objet(db: Session, detail_objet_id: int):
-    return db.query(DetailObjet).get(detail_objet_id)
+    detail_objet = db.query(DetailObjet).get(detail_objet_id)
+    if not detail_objet:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="DetailObjet not found")
+    return detail_objet
+
+
+def get_all_detail_objets(db: Session):
+    return db.query(DetailObjet).all()
 
 
 def create_detail_objet(db: Session, detail_objet: DetailObjet):
@@ -13,8 +21,12 @@ def create_detail_objet(db: Session, detail_objet: DetailObjet):
     return detail_objet
 
 
-def update_detail_objet(db: Session, detail_objet: DetailObjet, updated_detail_objet: DetailObjet):
-    for attr, value in updated_detail_objet.dict().items():
+def update_detail_objet(db: Session, detail_objet_id: int, updated_detail_objet: DetailObjet):
+    detail_objet = db.query(DetailObjet).get(detail_objet_id)
+    if not detail_objet:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="DetailObjet not found")
+    updated_data = updated_detail_objet.dict(exclude_unset=True)
+    for attr, value in updated_data.items():
         setattr(detail_objet, attr, value)
     db.commit()
     db.refresh(detail_objet)
@@ -22,6 +34,9 @@ def update_detail_objet(db: Session, detail_objet: DetailObjet, updated_detail_o
 
 
 def delete_detail_objet(db: Session, detail_objet: DetailObjet):
+    detail_objet = db.query(DetailObjet).get(detail_objet)
+    if not detail_objet:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="DetailObjet not found")
     db.delete(detail_objet)
     db.commit()
     return {"message": "DetailObjet deleted"}
