+++ projet/controllers/enseigne_controller.py
@@ -1,9 +1,17 @@
 from sqlalchemy.orm import Session
+from fastapi import HTTPException, status
 from ..models import Enseigne
 
 
 def get_enseigne(db: Session, enseigne_id: int):
-    return db.query(Enseigne).get(enseigne_id)
+    enseigne = db.query(Enseigne).get(enseigne_id)
+    if not enseigne:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Enseigne not found")
+    return enseigne
+
+
+def get_all_enseignes(db: Session):
+    return db.query(Enseigne).all()
 
 
 def create_enseigne(db: Session, enseigne: Enseigne):
@@ -13,15 +21,22 @@ def create_enseigne(db: Session, enseigne: Enseigne):
     return enseigne
 
 
-def update_enseigne(db: Session, enseigne: Enseigne, updated_enseigne: Enseigne):
-    for attr, value in updated_enseigne.dict().items():
+def update_enseigne(db: Session, enseigne_id: int, updated_enseigne: Enseigne):
+    enseigne = db.query(Enseigne).get(enseigne_id)
+    if not enseigne:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Enseigne not found")
+    updated_data = updated_enseigne.dict(exclude_unset=True)
+    for attr, value in updated_data.items():
         setattr(enseigne, attr, value)
     db.commit()
     db.refresh(enseigne)
     return enseigne
 
 
-def delete_enseigne(db: Session, enseigne: Enseigne):
+def delete_enseigne(db: Session, enseigne_id: Enseigne):
+    enseigne = db.query(Enseigne).get(enseigne_id)
+    if not enseigne:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Enseigne not found")
     db.delete(enseigne)
     db.commit()
     return {"message": "Enseigne deleted"}
