+++ projet/controllers/objet_cond_controller.py
@@ -1,9 +1,17 @@
 from sqlalchemy.orm import Session
+from fastapi import HTTPException, status
 from ..models import ObjetCond
 
 
 def get_objet_cond(db: Session, objet_cond_id: int):
-    return db.query(ObjetCond).get(objet_cond_id)
+    objet_cond = db.query(ObjetCond).get(objet_cond_id)
+    if not objet_cond:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="ObjetCond not found")
+    return objet_cond
+
+
+def get_all_objet_cond(db: Session):
+    return db.query(ObjetCond).all()
 
 
 def create_objet_cond(db: Session, objet_cond: ObjetCond):
@@ -13,15 +21,22 @@ def create_objet_cond(db: Session, objet_cond: ObjetCond):
     return objet_cond
 
 
-def update_objet_cond(db: Session, objet_cond: ObjetCond, updated_objet_cond: ObjetCond):
-    for attr, value in updated_objet_cond.dict().items():
+def update_objet_cond(db: Session, objet_cond_id: ObjetCond, updated_objet_cond: ObjetCond):
+    objet_cond = db.query(ObjetCond).get(objet_cond_id)
+    if not objet_cond:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="ObjetCond not found")
+    updated_data = updated_objet_cond.dict(exclude_unset=True)
+    for attr, value in updated_data.items():
         setattr(objet_cond, attr, value)
     db.commit()
     db.refresh(objet_cond)
     return objet_cond
 
 
-def delete_objet_cond(db: Session, objet_cond: ObjetCond):
+def delete_objet_cond(db: Session, objet_cond_id: int):
+    objet_cond = db.query(ObjetCond).get(objet_cond_id)
+    if not objet_cond:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="ObjetCond not found")
     db.delete(objet_cond)
     db.commit()
     return {"message": "ObjetCond deleted"}
