+++ projet/controllers/poids_controller.py
@@ -1,9 +1,17 @@
 from sqlalchemy.orm import Session
+from fastapi import HTTPException, status
 from ..models import Poids
 
 
 def get_poids(db: Session, poids_id: int):
-    return db.query(Poids).get(poids_id)
+    poids = db.query(Poids).get(poids_id)
+    if not poids:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Poids not found")
+    return poids
+
+
+def get_all_poids(db: Session):
+    return db.query(Poids).all()
 
 
 def create_poids(db: Session, poids: Poids):
@@ -13,15 +21,22 @@ def create_poids(db: Session, poids: Poids):
     return poids
 
 
-def update_poids(db: Session, poids: Poids, updated_poids: Poids):
-    for attr, value in updated_poids.dict().items():
+def update_poids(db: Session, poids_id: int, updated_poids: Poids):
+    poids = db.query(Poids).get(poids_id)
+    if not poids:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Poids not found")
+    updated_data = updated_poids.dict(exclude_unset=True)
+    for attr, value in updated_data.items():
         setattr(poids, attr, value)
     db.commit()
     db.refresh(poids)
     return poids
 
 
-def delete_poids(db: Session, poids: Poids):
+def delete_poids(db: Session, poids_id: Poids):
+    poids = db.query(Poids).get(poids_id)
+    if not poids:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Poids not found")
     db.delete(poids)
     db.commit()
     return {"message": "Poids deleted"}
