+++ projet/controllers/role_controller.py
@@ -1,9 +1,17 @@
 from sqlalchemy.orm import Session
+from fastapi import HTTPException, status
 from ..models import Role
 
 
 def get_role(db: Session, role_id: int):
-    return db.query(Role).get(role_id)
+    role = db.query(Role).get(role_id)
+    if not role:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Role not found")
+    return role
+
+
+def get_all_roles(db: Session):
+    return db.query(Role).all()
 
 
 def create_role(db: Session, role: Role):
@@ -13,15 +21,22 @@ def create_role(db: Session, role: Role):
     return role
 
 
-def update_role(db: Session, role: Role, updated_role: Role):
-    for attr, value in updated_role.dict().items():
+def update_role(db: Session, role_id: int, updated_role: Role):
+    role = db.query(Role).get(role_id)
+    if not role:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Role not found")
+    updated_data = updated_role.dict(exclude_unset=True)
+    for attr, value in updated_data.items():
         setattr(role, attr, value)
     db.commit()
     db.refresh(role)
     return role
 
 
-def delete_role(db: Session, role: Role):
+def delete_role(db: Session, role_id: int):
+    role = db.query(Role).get(role_id)
+    if not role:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Role not found")
     db.delete(role)
     db.commit()
     return {"message": "Role deleted"}
