+++ projet/controllers/vignette_controller.py
@@ -1,9 +1,17 @@
 from sqlalchemy.orm import Session
+from fastapi import HTTPException, status
 from ..models import Vignette
 
 
 def get_vignette(db: Session, vignette_id: int):
-    return db.query(Vignette).get(vignette_id)
+    vignette = db.query(Vignette).get(vignette_id)
+    if not vignette:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Vignette not found")
+    return vignette
+
+
+def get_all_vignettes(db: Session):
+    return db.query(Vignette).all()
 
 
 def create_vignette(db: Session, vignette: Vignette):
@@ -13,15 +21,22 @@ def create_vignette(db: Session, vignette: Vignette):
     return vignette
 
 
-def update_vignette(db: Session, vignette: Vignette, updated_vignette: Vignette):
-    for attr, value in updated_vignette.dict().items():
+def update_vignette(db: Session, vignette_id: int, updated_vignette: Vignette):
+    vignette = db.query(Vignette).get(vignette_id)
+    if not vignette:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Vignette not found")
+    updated_data = updated_vignette.dict(exclude_unset=True)
+    for attr, value in updated_data.items():
         setattr(vignette, attr, value)
     db.commit()
     db.refresh(vignette)
     return vignette
 
 
-def delete_vignette(db: Session, vignette: Vignette):
+def delete_vignette(db: Session, vignette_id: int):
+    vignette = db.query(Vignette).get(vignette_id)
+    if not vignette:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Vignette not found")
     db.delete(vignette)
     db.commit()
     return {"message": "Vignette deleted"}
